requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure Functions Python library (new programming model)
azure-functions==1.20.0

# MSAL for AAD client-credentials auth
msal>=1.24.0,<2.0

# HTTP calls to Dataverse Web API
requests>=2.31.0,<3.0


tzdata>=2024.1
